refactor: Move find_available_port() to separate port_allocation submodule
